+++ pytest/oar/__init__.py
@@ -3,7 +3,7 @@ import logging
 from .client import Client
 from .config import EnvConfig
 from .consts import Outcome, Analysis, Resolution
+from .models import Test, TestQuery, TestQueryResult
 from .report import Report
-from .result import Test, TestQuery, TestQueryResult
 
 logging.basicConfig(level=logging.INFO, format="%(message)s")
